official_writeup/misc-llm/game/requirements.txt
sanic
openai
pyOpenSSL